+++ src/MockData.jsx
@@ -1,0 +1,42 @@
+// ! REMOVE THIS LATER, FOR MOCKUP PURPOSES
+
+const MockData = [
+    {   
+        id: "01", 
+        name: "Jane Doe",
+        avatarUrl: null,
+        email: "jane@example.com",
+        password: "Jane_Doe123",
+        status: "Admin"
+    },
+    {   
+        id: "02", 
+        name: "James Doeker",
+        avatarUrl: null,
+        email: "james@example.com",
+        password: "Jamie122333",
+        status: "Non-Admin"
+    },
+    {   
+        id: "03", 
+        name: "Walahdalah",
+        avatarUrl: null,
+        email: "Walahdalah@example.com",
+        password: "Walah_LAH_LAH",
+        status: "Admin"
+    },
+];
+
+const currentUser = {
+    id: "04", 
+    name: "Developer",
+    avatarUrl: null,
+    email: "developer@example.com",
+    password: "Developer123",
+    status: "Admin"
+};
+
+export default MockData;
+export {
+    currentUser,
+}